@nllong example measure using extension gem
OsLib_AedgMeasures should either be added to extension gem or I should have rake task here to maintain it (although it isn't activly developed right now). I thought about idea of having measure gems like this and model articulation having their own library, but that would only work if those were installed with OpenStuido, and if we start  doign that, before you know it all of the measure will be installed with core OpenStudio which I don't think is what we want.